analyzer: split closure.h into sensible pieces.
